Extracting a class for FileSetDerivativesService
The purpose of this refactor is to build towards a more perfect world
by:

1) Reducing the number of methods added to a model, thus narrowing its
   interface (both public and private)
2) Exposing a clear collaborating class with a narrow purpose
3) To show a methodology for not abusing ActiveSupport::Concern
4) To be explicit in how a module is dependent on the class it mixes
   into, as well as what methods are important for that class to get
   from the module - see the delegate methods for how the service and
   FileSet are to interact.

Related to projecthydra/hydra-works#315
